+++ CHANGELOG.md
@@ -1,16 +1,17 @@
 # Change Log
 
-<!-- latest_release 19.0.4 -->
-## [v19.0.4](https://github.com/chef/ohai/tree/v19.0.4) (2024-10-22)
+<!-- latest_release 19.0.5 -->
+## [v19.0.5](https://github.com/chef/ohai/tree/v19.0.5) (2024-10-29)
 
 #### Merged Pull Requests
-- FreeBSD: Permit underscores in interface names [#1837](https://github.com/chef/ohai/pull/1837) ([feld](https://github.com/feld))
+- Unit tests for freebsd network plugin; plus fixes [#1841](https://github.com/chef/ohai/pull/1841) ([jaymzh](https://github.com/jaymzh))
 <!-- latest_release -->
 
 <!-- release_rollup since=18.1.3 -->
 ### Changes not yet released to rubygems.org
 
 #### Merged Pull Requests
+- Unit tests for freebsd network plugin; plus fixes [#1841](https://github.com/chef/ohai/pull/1841) ([jaymzh](https://github.com/jaymzh)) <!-- 19.0.5 -->
 - FreeBSD: Permit underscores in interface names [#1837](https://github.com/chef/ohai/pull/1837) ([feld](https://github.com/feld)) <!-- 19.0.4 -->
 - Migrate from Chefstyle to Cookstyle gem [#1833](https://github.com/chef/ohai/pull/1833) ([dafyddcrosby](https://github.com/dafyddcrosby)) <!-- 19.0.3 -->
 - [#1830] Check that sysfs directory exists before reading [#1831](https://github.com/chef/ohai/pull/1831) ([dafyddcrosby](https://github.com/dafyddcrosby)) <!-- 19.0.2 -->